+++ src/DotNetWorker/Configuration/ServiceCollectionExtensions.cs
@@ -1,4 +1,4 @@
-// Copyright (c) .NET Foundation. All rights reserved.
+﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 using System;
@@ -74,7 +74,7 @@ namespace Microsoft.Extensions.DependencyInjection
         internal static IServiceCollection RegisterDefaultConverters(this IServiceCollection services)
         {
             return services.AddSingleton<IConverter, OutputBindingConverter>()
-                           .AddSingleton<IConverter, ExactMatchConverter>()
+                           .AddSingleton<IConverter, TypeConverter>()
                            .AddSingleton<IConverter, JsonPocoConverter>();
         }