+++ src/DotNetWorker/Converters/TypeConverter.cs
@@ -1,13 +1,17 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
+using System;
+
 namespace Microsoft.Azure.Functions.Worker.Converters
 {
-    internal class ExactMatchConverter : IConverter
+    internal class TypeConverter : IConverter
     {
         public bool TryConvert(ConverterContext context, out object? target)
         {
-            if (context.Source?.GetType() == context.Parameter.Type)
+            Type? sourceType = context.Source?.GetType();
+            if (sourceType is not null &&
+                context.Parameter.Type.IsAssignableFrom(sourceType))
             {
                 target = context.Source;
                 return true;