+++ src/DotNetWorker/DefaultHostRequestHandler.cs
@@ -1,4 +1,4 @@
-// Copyright (c) .NET Foundation. All rights reserved.
+﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 using System;
@@ -20,10 +20,18 @@ namespace Microsoft.Azure.Functions.Worker
 
         public Task<WorkerInitResponse> InitializeWorkerAsync(WorkerInitRequest request)
         {
-            var response = new WorkerInitResponse
+            var response = new WorkerInitResponse()
             {
                 Result = new StatusResult { Status = Status.Success }
             };
+
+
+            response.Capabilities.Add("RpcHttpBodyOnly", bool.TrueString);
+            response.Capabilities.Add("RawHttpBodyBytes", bool.TrueString);
+            response.Capabilities.Add("RpcHttpTriggerMetadataRemoved", bool.TrueString);
+            response.Capabilities.Add("UseNullableValueDictionaryForHttp", bool.TrueString);
+
+            response.WorkerVersion = typeof(DefaultHostRequestHandler).Assembly.GetName().Version?.ToString();
 
             return Task.FromResult(response);
         }