+++ src/DotNetWorker/Http/SameSite.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License. See License.txt in the project root for license information.
+
+namespace Microsoft.Azure.Functions.Worker
+{
+    public enum SameSite
+    {
+        None = 0,
+        Lax = 1,
+        Strict = 2,
+        ExplicitNone = 3,
+    }
+}