+++ release_notes.md
@@ -2,3 +2,4 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR/#issue)
 -->
+- Enhancements to the HttpRequestData type and APIs (#120)